Refactor naming and organization
