+++ src/qris_dockwidget.ui
