+++ src/ui/add_detrended_dialog.ui
