+++ src/ui/add_layer_dialog.ui
